Fix to prevent FreeBSD symbol tables from being remapped. With this fix both global and stack frame local variables can now be inspected by name.
Details:
The r_debug link map is provided by the loader to communicate shared
library load paths and addresses to a debugger. Linux uses an empty
path to indicate the executable which was already being check. FreeBSD
was providing the path to the executable which was interpretted as a
shared library and caused the remap to occur. Both an empty path and
the executable path are now checked before adding shared libraries.